backend/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-socketio==5.3.6
scikit-learn>=1.6.0
numpy>=1.26.0
python-dotenv==1.0.0
eventlet>=0.35.2
gunicorn>=21.2.0
psycopg[binary]
cloudinary
